Rename package and return SessionDTO
